Implemented "midimessage" event dispatching on noteOn/noteOff
When the visual keyboard of the synth gets clicked, a midi message
event gets created and dispatched on window.onmidimessage (on window
because there is no prototype to dispatch on currently). This allows
third party software to use the virtual keyboard clicks as MIDI input.
This impl. is using the data structure of the WebMIDIApi shim to be
inter-operable and standard conform. Thank you   @cwilso — WebMIDIApi
is awesome!